# Conflicts:
#	cupy/random/__init__.py
#	cupy/random/distributions.py
#	cupy/random/generator.py
#	docs/source/reference/random.rst
#	tests/cupy_tests/random_tests/test_distributions.py
#	tests/cupy_tests/random_tests/test_generator.py

+++ cupy/core/include/cupy/carray.cuh
@@ -122,7 +122,7 @@ public:
     } else if (x.iszero()) {
       ret_raw_.x = (y_raw_.x & 0x8000u) + 1;
     } else if (!(x_raw_.x & 0x8000u)) {
-      if (x_raw_.x > y_raw_.x) {
+      if (static_cast<signed short>(x_raw_.x) > static_cast<signed short>(y_raw_.x)) {
         ret_raw_.x = x_raw_.x - 1;
       } else {
         ret_raw_.x = x_raw_.x + 1;
